refactor(trusty-mpm): extract launch-on-main into its own module + collision-helper coverage (#3455)
Addresses code-critic WARN on PR #3781:

1. Doc-pointer lint (RED): spawn_managed_on_main's `Test:` pointer cited
   `spawn_managed_on_main_warns_on_concurrent_main_checkout_session`, which
   did not exist. Added it — a pure unit test over a new extracted helper
   `has_concurrent_main_checkout_session(existing, local_path) -> Option<&Record>`
   that returns the FIRST Active session sharing the exact cwd (asserts it
   fires on an Active same-cwd record and is None for a different path, a
   non-Active record, or an empty set). Pointer now names the real test.

2. 500-line file-size cap (RED): lifecycle.rs had grown to 959 SLOC vs its
   frozen 840 budget. Extracted worktree_enabled_for_origin,
   has_concurrent_main_checkout_session, and spawn_managed_on_main into a new
   sibling module launch_on_main.rs (matching the inproject.rs split). The
   shared spawn helpers they reuse (write_task_md/prepare_inproject_session/
   front_gate_or_escalate/resolve_gh_env) become pub(super) and are called
   back into. lifecycle.rs is back to 830 SLOC (< 840); launch_on_main.rs is
   146 SLOC (< 500).

3. Diagnosability: the concurrent-main-checkout warn! now uses the .find()
   helper and logs the colliding session's id + tmux name so an operator can
   identify which other session shares the checkout.
