+++ crates/trusty-mpm/src/daemon/managed_routes/mod.rs
@@ -33,6 +33,7 @@ mod fleet;
 pub mod front_gate;
 pub mod inproject;
 pub mod inproject_hygiene;
+mod launch_on_main;
 mod lifecycle;
 mod mcp_spawn_gate;
 mod project_registry_routes;
